+++ docs/Whats-New.md
@@ -224,6 +224,7 @@ Vanilla fixes:
 - Fix a bug where you could sometimes get extra crew to exit a building that was being sold and was destroying/undeploying (by ZivDero)
 - Fix a bug where if the player loaded a saved game, the score screen timer would report the time since the saved game was loaded, instead of since when the scenario was first started (by ZivDero)
 - Fix a bug where units could gain veterancy by killing allies (by ZivDero)
+- Fix a bug where AI players would send teams to attack their allies (by ZivDero)
 
 </details>